+++ app/components/Snaps/SnapUICard/SnapUICard.tsx
@@ -7,7 +7,6 @@ import Text, {
 import { SnapUIImage } from '../SnapUIImage/SnapUIImage';
 import {
   FlexDirection,
-  TextAlign,
   JustifyContent,
   AlignItems,
 } from '../../UI/Box/box.types';
@@ -54,12 +53,22 @@ export const SnapUICard: FunctionComponent<SnapUICardProps> = ({
         )}
       </Box>
     </Box>
-    <Box flexDirection={FlexDirection.Column} textAlign={TextAlign.right}>
-      <Text variant={TextVariant.BodyMDMedium} ellipsizeMode="tail">
+    <Box flexDirection={FlexDirection.Column}>
+      <Text
+        variant={TextVariant.BodyMDMedium}
+        ellipsizeMode="tail"
+        // eslint-disable-next-line react-native/no-inline-styles
+        style={{ textAlign: 'right' }}
+      >
         {value}
       </Text>
       {extra && (
-        <Text color={TextColor.Alternative} ellipsizeMode="tail">
+        <Text
+          color={TextColor.Alternative}
+          ellipsizeMode="tail"
+          // eslint-disable-next-line react-native/no-inline-styles
+          style={{ textAlign: 'right' }}
+        >
           {extra}
         </Text>
       )}